+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2
+
+- Add themes
+- Add repeat pixel icon
+- Fix issues where filtering and unfiltering would pause the current playing song
+- Fix issue where navigating to other activities will pause the song
+
 ## 1
 
 - Add playlist view